# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InvestPro Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List user withdrawals
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /withdrawals/{id}:
    get:
      operationId: getWithdrawal
      tags: [withdrawals]
      summary: Get withdrawal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"

  # Referrals
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List support tickets
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
  /tickets/{id}/reply:
    post:
      operationId: replyTicket
      tags: [tickets]
      summary: Reply to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyInput"
      responses:
        "201":
          description: Reply added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"

  # Admin
...
components:
  schemas:
...
    Withdrawal:
      type: object
      required: [id, amount, fee, netAmount, status, walletAddress, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        walletAddress:
          type: string
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
    WithdrawalInput:
      type: object
      required: [amount, walletAddress]
      properties:
        amount:
          type: number
        walletAddress:
          type: string

    # Referrals
...
    Ticket:
      type: object
      required: [id, subject, category, status, createdAt, lastReplyAt]
      properties:
        id:
          type: integer
        subject:
          type: string
        category:
          type: string
          enum: [deposit, withdrawal, kyc, technical, other]
        status:
          type: string
          enum: [open, replied, closed]
        messageCount:
          type: integer
        createdAt:
          type: string
        lastReplyAt:
          type: string
    TicketDetail:
      type: object
      required: [id, subject, category, status, createdAt, messages]
      properties:
        id:
          type: integer
        subject:
          type: string
        category:
          type: string
        status:
          type: string
        createdAt:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/TicketMessage"
    TicketMessage:
      type: object
      required: [id, message, isAdmin, createdAt]
      properties:
        id:
          type: integer
        message:
          type: string
        isAdmin:
          type: boolean
        authorName:
          type: string
        createdAt:
          type: string
    TicketInput:
      type: object
      required: [subject, category, message]
      properties:
        subject:
          type: string
        category:
          type: string
        message:
          type: string
    TicketReplyInput:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Admin
...