# lib/api-zod/src/generated/types/ticketDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InvestPro Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TicketMessage } from './ticketMessage';

export interface TicketDetail {
  id: number;
  subject: string;
  category: string;
  status: string;
  createdAt: string;
  messages: TicketMessage[];
}